Duplicate code removed
