+++ app/src/main/java/org/piramalswasthya/sakhi/ui/home_activity/village_level_forms/pulse_polio/PulsePolioCampaignListViewModel.kt
@@ -10,6 +10,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import org.piramalswasthya.sakhi.repositories.VLFRepo
+import org.piramalswasthya.sakhi.utils.CampaignDateUtil
 import org.piramalswasthya.sakhi.utils.HelperUtil
 import javax.inject.Inject
 @RequiresApi(Build.VERSION_CODES.O)
@@ -24,7 +25,7 @@ class PulsePolioCampaignListViewModel @Inject constructor(
             list.filter {
                 try {
                     val date = it.campaignDate ?: return@filter false
-                    getYearFromDate(date) == currentYear
+                    CampaignDateUtil.getYearFromDate(date) == currentYear
                 } catch (e: Exception) {
                     false
                 }
@@ -56,7 +57,7 @@ class PulsePolioCampaignListViewModel @Inject constructor(
                 val list = vlfRepo.getAllPulsePolioCampaigns()
                 val currentYearRecords = list.filter {
                     val date = it.campaignDate ?: return@filter false
-                    getYearFromDate(date) == currentYear
+                    CampaignDateUtil.getYearFromDate(date) == currentYear
                 }
 
                 if (currentYearRecords.size >= 2) {
@@ -66,7 +67,7 @@ class PulsePolioCampaignListViewModel @Inject constructor(
 
                 val allCampaignDates = list.mapNotNull { it.campaignDate }
                 val mostRecentCampaignDate = allCampaignDates.maxByOrNull { dateStr ->
-                    parseDateToLocalDate(dateStr)?.toEpochDay() ?: Long.MIN_VALUE
+                    CampaignDateUtil.parseDateToLocalDate(dateStr)?.toEpochDay() ?: Long.MIN_VALUE
                 }
 
                 if (mostRecentCampaignDate == null) {
@@ -74,7 +75,7 @@ class PulsePolioCampaignListViewModel @Inject constructor(
                     return@launch
                 }
 
-                val canAdd = isSixMonthsCompleted(mostRecentCampaignDate)
+                val canAdd = CampaignDateUtil.isMonthsCompleted(mostRecentCampaignDate, 6)
                 _isCampaignAlreadyAdded.postValue(!canAdd)
             } catch (e: Exception) {
                 _isCampaignAlreadyAdded.postValue(false)
@@ -82,69 +83,4 @@ class PulsePolioCampaignListViewModel @Inject constructor(
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun parseDateToLocalDate(dateStr: String): java.time.LocalDate? {
-        return try {
-            val dateFormats = listOf(
-                "yyyy-MM-dd",
-                "dd-MM-yyyy",
-                "dd/MM/yyyy",
-                "yyyy/MM/dd"
-            )
-            
-            for (format in dateFormats) {
-                try {
-                    val formatter = java.time.format.DateTimeFormatter.ofPattern(format)
-                    return java.time.LocalDate.parse(dateStr, formatter)
-                } catch (e: Exception) {
-                }
-            }
-            null
-        } catch (e: Exception) {
-            null
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun getYearFromDate(dateStr: String): Int {
-        return try {
-            val date = parseDateToLocalDate(dateStr)
-            if (date != null) {
-                return date.year
-            }
-
-            if (dateStr.length >= 4 && (dateStr[4] == '-' || dateStr[4] == '/')) {
-                dateStr.take(4).toInt()
-            } 
-            else if (dateStr.length >= 10) {
-                val lastSeparatorIndex = maxOf(
-                    dateStr.lastIndexOf('-'),
-                    dateStr.lastIndexOf('/')
-                )
-                if (lastSeparatorIndex > 0 && lastSeparatorIndex < dateStr.length - 4) {
-                    dateStr.substring(lastSeparatorIndex + 1).toInt()
-                } else {
-                    0
-                }
-            } else {
-                0
-            }
-        } catch (e: Exception) {
-            0
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun isSixMonthsCompleted(lastDate: String): Boolean {
-        return try {
-            val last = parseDateToLocalDate(lastDate) ?: return false
-
-            val nextAllowed = last.plusMonths(6)
-            val today = java.time.LocalDate.now()
-
-            today.isAfter(nextAllowed) || today.isEqual(nextAllowed)
-        } catch (e: Exception) {
-            false
-        }
-    }
 }
